+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@ This is the offical release of vpinfe v1.1.8.
 
 ### What's New
 - Centrilized JS sound API now in vpinfe-core.js.  No need for themes to implement their own audio playback.
+- Added play metrics.  Number of times a table was played (StartCount) and the minutes played (RunTime).
+- Ability to change the title of the table and set an AltVPSID (aka a table specifc id).
 
 ### Fixes
 None